# Speedup `filter_bytes` ~-20-40%, `filter_native` low selectivity (~-37%)