# Which issue does this PR close?

<!--
We generally require a GitHub issue to be filed for all bug fixes and enhancements and this helps us generate change logs for our releases. You can link an issue to this PR using the GitHub syntax. For example `Closes #123` indicates that this PR will close issue #123.
-->

Closes #7465

* We can precalculate capacity for the 
* Use `Vec` api for `filter_native` to generate some faster code.

# Rationale for this change
 
<!--
Why are you proposing this change? If this is already explained clearly in the issue then this section is not needed.
Explaining clearly why changes are proposed helps reviewers understand your changes and offer better suggestions for fixes.
-->

<details>

```
filter context string (kept 1/2)
                        time:   [494.54 µs 496.78 µs 499.12 µs]
                        change: [-4.6353% -2.8403% -1.1083%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 15 outliers among 100 measurements (15.00%)
  2 (2.00%) low severe
  2 (2.00%) low mild
  5 (5.00%) high mild
  6 (6.00%) high severe

filter context string high selectivity (kept 1023/1024)
                        time:   [655.05 µs 657.74 µs 660.21 µs]
                        change: [-42.752% -41.103% -39.592%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 24 outliers among 100 measurements (24.00%)
  5 (5.00%) low severe
  8 (8.00%) low mild
  4 (4.00%) high mild
  7 (7.00%) high severe

filter context string low selectivity (kept 1/1024)
                        time:   [616.29 ns 617.36 ns 618.40 ns]
                        change: [-26.996% -26.638% -26.282%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 11 outliers among 100 measurements (11.00%)
  6 (6.00%) high mild
  5 (5.00%) high severe


filter context i32 low selectivity (kept 1/1024)
                        time:   [139.28 ns 139.80 ns 140.41 ns]
                        change: [-37.655% -37.164% -36.612%] (p = 0.00 < 0.05)
                        Performance has improved.

```

# What changes are included in this PR?

<!--
There is no need to duplicate the description in the issue here but it is sometimes worth providing a summary of the individual changes in this PR.
-->

# Are there any user-facing changes?


<!--
If there are user-facing changes then we may require documentation to be updated before approving the PR.
-->

<!---
If there are any breaking changes to public APIs, please call them out.
-->