@alamb 2025-05-05
🤖 `./gh_compare_arrow.sh` [Benchmark Script](https://github.com/alamb/datafusion-benchmarking/blob/main/gh_compare_arrow.sh) Running
Linux aal-dev 6.11.0-1013-gcp #13~24.04.1-Ubuntu SMP Wed Apr  2 16:34:16 UTC 2025 x86_64 x86_64 x86_64 GNU/Linux
Comparing speedup_filters (816bab22546bebe865f1be855f6d239673f365ed) to 4491b170a7595e764de961bc1537197db0df9f7f [diff](https://github.com/apache/arrow-rs/compare/4491b170a7595e764de961bc1537197db0df9f7f..816bab22546bebe865f1be855f6d239673f365ed)
BENCH_NAME=filter_kernels
BENCH_COMMAND=cargo bench --all-features --bench filter_kernels 
BENCH_FILTER=
BENCH_BRANCH_NAME=speedup_filters
Results will be posted here when complete